Auth login fix # 3: Cover

+++ QR-Line/asset/js/navbar.js
@@ -9,8 +9,28 @@ if (localStorage.getItem("user_id")) {
 
 // Add event listener for the logout button
 document.getElementById("btnLogOut").addEventListener("click", function() {
-    localStorage.clear();
-    document.getElementById("divbtnLO").style.display = "none";
-    document.getElementById("btnLogOut").style.display = "none";
-    window.location.href = "./../auth/login.php";
+    $.ajax({
+        url: "./../api/api_authenticate.php",
+        type: "POST",
+        data: JSON.stringify({
+            username: "OK",
+            password: "OK",
+            auth_method: "logout"
+        }),
+        contentType: "application/json",
+        dataType: "json",
+        success: function(response) {
+            if (response.status === "success") {
+                localStorage.clear();
+                document.getElementById("divbtnLO").style.display = "none";
+                document.getElementById("btnLogOut").style.display = "none";
+                window.location.href = "./../auth/login.php";
+            } else {
+                console.log("Error:", response.message);
+            }
+        },
+        error: function(xhr, status, error) {
+            console.error("AJAX Error:", status, error);
+        }
+    });
 });
